Avoid flaky test using conditional wait
