+++ package.json
@@ -66,6 +66,7 @@
 		"tslint": "gulp tslint"
 	},
 	"dependencies": {
+		"@types/request": "^2.48.1",
 		"co": "^4.6.0",
 		"detect-port": "^1.3.0",
 		"expand-home-dir": "^0.0.3",
@@ -86,6 +87,7 @@
 		"gulp-tslint": "^8.1.3",
 		"mocha": "^5.2.0",
 		"mvn-artifact-download": "^4.0.0",
+		"p-wait-for": "^2.0.1",
 		"sinon": "^7.2.3",
 		"sinon-chai": "^3.3.0",
 		"tslint": "^5.12.1",
