+++ src/test/command.test.ts
@@ -1,12 +1,12 @@
 "use strict";
 
 import * as chai from "chai";
+import * as commandUtil from "./command.util";
+import * as pWaitFor from "p-wait-for";
 import * as sinon from "sinon";
 import * as sinonChai from "sinon-chai";
 import * as vscode from "vscode";
-import * as detect from 'detect-port';
 
-const request = require("request");
 const expect = chai.expect;
 chai.use(sinonChai);
 
@@ -18,7 +18,6 @@ describe("AtlasMap/Commands", function() {
 	let showWarningMessageSpy: sinon.SinonSpy;
 	let createOutputChannelSpy: sinon.SinonSpy;
 	let port: string;
-	const keyString: string = "Starting AtlasMap instance at port ";
 
 	before(async function() {
 		sandbox = sinon.createSandbox();
@@ -47,7 +46,7 @@ describe("AtlasMap/Commands", function() {
 		it("test command execution", function() {
 			// atlasmap has been started in the setup of the test suite already - just check the call count and determine the port
 			expect(executeCommandSpy.withArgs("atlasmap.start").calledOnce, "AtlasMap start command was not issued").to.be.true;
-			port = determineUsedPort();
+			port = commandUtil.determineUsedPort(showInformationMessageSpy);
 			expect(port, "Unable to determine used port for AtlasMap server").to.not.be.undefined;
 			expect(port, "Port for AtlasMap server seems to be NaN").to.not.be.NaN;
 			expect(createOutputChannelSpy.calledOnce);
@@ -73,120 +72,16 @@ describe("AtlasMap/Commands", function() {
 			expect(port, "Port for AtlasMap server seems to be NaN").to.not.be.NaN;
 
 			let url:string = "http://localhost:" + port;
-			const body = await getWebUI(url);
+			const body = await commandUtil.getWebUI(url);
 			expect(body, "Unexpected html response body").to.contain("AtlasMap");
-		});
-	});
 
-	describe("Stop AtlasMap Command Tests", function() {
-	
-		it("test stop command execution", async function() {
-			executeCommandSpy.resetHistory();
-			await vscode.commands.executeCommand("atlasmap.start");
-			expect(executeCommandSpy.withArgs("atlasmap.start").calledOnce, "AtlasMap start command was not issued").to.be.true;
-
-			port = determineUsedPort();
-			expect(port, "Unable to determine used port for AtlasMap server").to.not.be.undefined;
-			expect(port, "Port for AtlasMap server seems to be NaN").to.not.be.NaN;
-			expect(createOutputChannelSpy.calledOnce);
-
+			//Stop to avoid polluting other tests
 			await vscode.commands.executeCommand("atlasmap.stop");
-			expect(executeCommandSpy.withArgs("atlasmap.stop").calledOnce, "AtlasMap stop command was not issued").to.be.true;
-
-			await new Promise(resolve => setTimeout(resolve, 10000));
-
-			await retrieveFreeLocalPort(port)
-				.then( (_port) => {
-					expect("" + _port, "The port used to run AtlasMap seems to be still occupied. Stop command failed?").to.equal(port);
-				})
-				.catch( (err) => {
-					expect(err).to.be.null;
-				});
-		});
-
-		it("test stopping without AtlasMap running", async function() {
-			// make sure no atlasmap is running
-			await vscode.commands.executeCommand("atlasmap.stop");
-			await new Promise(resolve => setTimeout(resolve, 10000));
-			showWarningMessageSpy.resetHistory();
-
-			// try to stop the not running atlasmap
-			await vscode.commands.executeCommand("atlasmap.stop");
-			expect(showWarningMessageSpy.getCalls()[showWarningMessageSpy.callCount-1].args[0], "No warning message found that no AtlasMap has been detected!").to.equal("Unable to locate running AtlasMap instance");
-		});
-	});
-
-	describe("Combined Open/Stop/Open tests", function() {
-	
-		it("test output channel recreation", async function() {
-			executeCommandSpy.resetHistory();
-			await vscode.commands.executeCommand("atlasmap.start");
-			await new Promise(resolve => setTimeout(resolve, 15000));
-			expect(executeCommandSpy.withArgs("atlasmap.start").calledOnce, "AtlasMap start command was not issued").to.be.true;
-
-			port = determineUsedPort();
-			expect(port, "Unable to determine used port for AtlasMap server").to.not.be.undefined;
-			expect(port, "Port for AtlasMap server seems to be NaN").to.not.be.NaN;
-			expect(createOutputChannelSpy.withArgs("AtlasMap Server").calledOnce).to.be.true;
-
-			await vscode.commands.executeCommand("atlasmap.stop");
+			await pWaitFor(() => {
+				return showInformationMessageSpy.withArgs("Stopped AtlasMap instance at port " + port).calledOnce;
+			}, 500, 200);
 			expect(executeCommandSpy.withArgs("atlasmap.stop").calledOnce, "AtlasMap stop command was not issued").to.be.true;
 			await new Promise(resolve => setTimeout(resolve, 10000));
-			await vscode.commands.executeCommand("atlasmap.start");
-			await new Promise(resolve => setTimeout(resolve, 15000));
-			expect(executeCommandSpy.withArgs("atlasmap.start").calledTwice, "AtlasMap start command was not issued").to.be.true;
-
-			port = determineUsedPort();
-			expect(port, "Unable to determine used port for AtlasMap server").to.not.be.undefined;
-			expect(port, "Port for AtlasMap server seems to be NaN").to.not.be.NaN;
-			expect(createOutputChannelSpy.withArgs("AtlasMap Server").calledTwice).to.be.true;
-
-			let url:string = "http://localhost:" + port;
-			const body = await getWebUI(url);
-			expect(body, "Unexpected html response body").to.contain("AtlasMap");
 		});
 	});
-
-	function retrieveFreeLocalPort(testport: string): Promise<string> {
-		return new Promise((resolve, reject) => {
-			detect(testport)
-				.then(_port => {
-					resolve(_port);
-				})
-				.catch(err => {
-					reject(err);
-				});
-		});
-	}
-
-	function getWebUI(url: string) {
-		return new Promise((resolve, reject) => {
-			request(url, (error: any, response: any, body: any) => {
-				if (error) reject(error);
-				if (!response || response.statusCode != 200) {
-					reject('Invalid response');
-				} else {
-					resolve(body);
-				}
-			});
-		});
-	}
-
-	function determineUsedPort(): string {
-		let port: string;
-		for (var callIdx = 0; callIdx < showInformationMessageSpy.callCount; callIdx++) {
-			let cal = showInformationMessageSpy.getCall(callIdx);
-			for (var argIdx = 0; argIdx < cal.args.length; argIdx++) {
-				var arg = cal.args[argIdx];
-				if (arg.startsWith(keyString)) {
-					port = arg.substring(keyString.length);
-					break;
-				}
-			}
-			if (port !== undefined) {
-				break;
-			}
-		}
-		return port;
-	}
 });
